It's been a while since a clean commit.

+++ runables/run_aliscore.sh
@@ -1,7 +1,15 @@
 #! /bin/bash
 
-cp /fslgroup/fslg_BybeeLab/scripts/nick/aliscore/Aliscore.02.2.pl ./
-cp /fslgroup/fslg_BybeeLab/scripts/nick/aliscore/Aliscore_module.pm ./
+
+if ! [ -e Aliscore.02.2.pl ]
+then
+	cp /fslgroup/fslg_BybeeLab/scripts/nick/aliscore/Aliscore.02.2.pl ./
+fi
+
+if ! [ -e Aliscore_module.pm ]
+then
+	cp /fslgroup/fslg_BybeeLab/scripts/nick/aliscore/Aliscore_module.pm ./
+fi
 
 for file in *.aln
 do
@@ -10,6 +18,6 @@ do
 	echo '#SBATCH --time=5:00 --ntasks=1 --nodes=1 --mem-per-cpu=8G -J Ali' >> $batch
 	echo '' >> $batch
 	echo "perl Aliscore.02.2.pl -i $file" >> $batch
-	sbatch $batch
+#	sbatch $batch
 done
